PUT /perms/users/{id} and PUT /perms/permissions/{id} to return 204, not 200
Currently
PUT /perms/users/{id} and
PUT /perms/permissions/{id}
return 200 and the input entity on success.

The client does not need the input entity.
The API doesn't return the metadata property with the new data
created by the database trigger.
Most other FOLIO PUT APIs do not return the entity and use
204 "No content" instead of the 200 HTTP status code.

This pull request changes from 200 to 204 HTTP status code.

Open question:
Is this a breaking change so that we need to bump the interface version
from 5.3 to 6.0 and the mod-permissions version from 5.11.3 to 6.0.0?