Do not assign context to dummy permissions MODPERMS-133 (#114)
Dummy permissions should not be assigned a context. They don't belong to the module where they happen to
be referred to.

They should be assigned a context when they are really defined.

Furthermore the collision check in savePerms is wrong for. It checks if context is null and then assumes
it's user-defined context. If a permission has Mutable undefined or TRUE – then it'a user-defined permission.