MODPERMS-156: Perms users request results in invalid SQL
POST /perms/users fails with an SQL error if no userId is
supplied.

To fix this userId is marked as required in permissionUser.json and
is converted into a uniqueIndex in schema.json.

This replaces the Java code that check for duplicate userId, that code
is removed.

The id is always the primary key, this has always a unique index.
The Java code that checks for duplicate id is not needed.
The non-unique index for id is not needed and is removed
from schema.json.

The id variable name incorrectly was userid and is renamed to id.